Implemented Place Order but getting stack Over Flow while clearing the Cart items

+++ src/main/java/com/ecom/ecommerceapis/Models/OrderItem.java
@@ -15,7 +15,7 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int productId;
+    private Long productId;
 
     private int quantity;
